refactor: 기능별 클래스 분리
