+++ docs/README.md
@@ -13,9 +13,9 @@
   - [x] 우승자가 여러명일 경우 ,를 사용하여 구분한다.
 ### 테스트 기능 요구사항
 - [x] JUnit5와 AssertJ를 이용하여 정리한 기능 목록 동작을 테스트 코드로 확인한다.
-- [ ] 입력 기능을 테스트한다.
-- [ ] 각 차수별 실행 결과 출력 기능을 테스트한다.
-- [ ] 최종 우승자 출력 기능을 테스트한다.
+- [x] 입력 기능을 테스트한다.
+- [x] 각 차수별 실행 결과 출력 기능을 테스트한다.
+- [x] 최종 우승자 출력 기능을 테스트한다.
 ### 추가 고려사항
 - [x] indent depth가 3이 넘지 않도록 한다.
 - [x] 함수가 한 가지 일만 하도록 최대한 작게 만든다.
